[refactor] dto패키지 위치 변경
