+++ src/main/java/com/cow/cow_mvc_practice/member/service/MemberService.java
@@ -2,7 +2,7 @@ package com.cow.cow_mvc_practice.member.service;
 
 import java.util.List;
 
-import com.cow.cow_mvc_practice.member.controller.dto.response.MemberResponse;
+import com.cow.cow_mvc_practice.member.dto.response.MemberResponse;
 
 public interface MemberService {
