+++ src/main/java/com/cow/cow_mvc_practice/post/dto/request/FindPostRequest.java
@@ -1,15 +1,20 @@
-package com.cow.cow_mvc_practice.post.controller.dto.request;
+package com.cow.cow_mvc_practice.post.dto.request;
 
 import com.cow.cow_mvc_practice.member.entity.Member;
 import com.cow.cow_mvc_practice.post.entity.Post;
 import lombok.Getter;
 
 @Getter
-public class DeletePostRequest {
-    Long memberId;
+public class FindPostRequest {
+    Long id;
+    String content;
+    String title;
 
     public Post toEntity(Member member) {
         return Post.builder()
+                .title(title)
+                .content(content)
+                .member(member)
                 .build();
     }
 }
