+++ src/main/java/com/cow/cow_mvc_practice/post/service/PostServiceImpl.java
@@ -3,10 +3,10 @@ package com.cow.cow_mvc_practice.post.service;
 import com.cow.cow_mvc_practice.comment.repository.CommentJPARepository;
 import com.cow.cow_mvc_practice.member.entity.Member;
 import com.cow.cow_mvc_practice.member.repository.MemberJPARepository;
-import com.cow.cow_mvc_practice.post.controller.dto.request.CreatePostRequest;
-import com.cow.cow_mvc_practice.post.controller.dto.request.DeletePostRequest;
-import com.cow.cow_mvc_practice.post.controller.dto.response.CreatePostResponse;
-import com.cow.cow_mvc_practice.post.controller.dto.response.FindPostResponse;
+import com.cow.cow_mvc_practice.post.dto.request.CreatePostRequest;
+import com.cow.cow_mvc_practice.post.dto.request.DeletePostRequest;
+import com.cow.cow_mvc_practice.post.dto.response.CreatePostResponse;
+import com.cow.cow_mvc_practice.post.dto.response.FindPostResponse;
 import com.cow.cow_mvc_practice.post.entity.Post;
 import com.cow.cow_mvc_practice.post.repository.PostJPARepository;
 import jakarta.persistence.EntityNotFoundException;
